[wired-input] Fix value property.
Setting the value property after the wired-input is created has
no effect since the element does not update.  Make `value` a
normal property and use it's value to update the value attribute
of the contained input in the render method.

[wired-checkbox] Allow changing checked state.

Setting the checked state of a wired-checkbox after it has been
created had no effect.  Make sure the check visiblity of the
check on initial draw and on update.